Updated: It is now possible to weight the attributes

+++ clustering/src/main/java/org/meyerlab/nopence/clustering/algorithms/hierarchical/terminateOptions/ClusterSizeTerminateOption.java
@@ -7,9 +7,9 @@ import org.meyerlab.nopence.clustering.algorithms.hierarchical.HierarchicalClust
  */
 public class ClusterSizeTerminateOption implements ITerminateOption {
 
-    private final int _minClusterSize;
+    private final double _minClusterSize;
 
-    public ClusterSizeTerminateOption(int minClusterSize) {
+    public ClusterSizeTerminateOption(double minClusterSize) {
         _minClusterSize = minClusterSize;
     }
